# fix: Tunning performance for sales documents with many itens 20+