--- a/frappe/public/js/frappe/model/model.js
+++ b/frappe/public/js/frappe/model/model.js
@@ -651,25 +651,26 @@ $.extend(frappe.model, {
 	},
 
 	round_floats_childs_in: function(arr, fieldnames) {
-		var cached_precision = [];
-		$.each(arr || [], function(k, item) {
-			if(!fieldnames) {
-				fieldnames = frappe.meta.get_fieldnames(item.doctype, item.parent,
+		var precision_map = [];
+		if (arr && arr.length > 0) {
+			var firstRow = arr[0];
+			if (!fieldnames) {
+				fieldnames = frappe.meta.get_fieldnames(firstRow.doctype, firstRow.parent,
 					{"fieldtype": ["in", ["Currency", "Float"]]});
 			}
 
-			for(var i=0, j=fieldnames.length; i < j; i++) {
-				var fieldname = fieldnames[i];
-				
-				if (k == 0) { // cache precision on first arr position, the precision will be the same for all rows.
-					// precision was cached because it's running many times, and tuning the loop to many itens on sales documents...
-					cached_precision[fieldname] = precision(fieldname, item);
-				}
+			fieldnames.forEach(fieldname => {
+				// precision was cached because it's running many times, and tuning the loop to many itens on sales documents...
+				precision_map[fieldname] = precision(fieldname, firstRow);
+			})
 
-				item[fieldname] = flt(item[fieldname], cached_precision[fieldname]);
-			}
-		});
-	},	
+			arr.forEach(row => {
+				fieldnames.forEach(fieldname => {
+					row[fieldname] = flt(row[fieldname], precision_map[fieldname]);
+				})
+			})			
+		};
+	},
 
 	validate_missing: function(doc, fieldname) {
 		if(!doc[fieldname]) {