@surajshetty3416 2021-11-23
@HarryPaulo Since the utility function is not generic enough I would recommend you to keep this function in ERPNext where it is used.

Thanks!